Setup webpack front end bundles
Use "npm run dev" to run the app on a local dev server, which auto refershes after changes are made to source code.
The dev server will not build any output files as it just stores them in memory.
Use "npm run build" to create the final output in /dist folder.
The contents of the dist folder is then ready to copy to the server directly as a complete build.
- Move all development files to src folder, including HTML and CSS
- Install webpack with required modules and pugins
- Setup webpack dev server
- src/index.html has styles and scripts injected automatically so no tags required
- Note: HistropediaJS is still being called manually with a <script> tag.
  This will be reconciled into the build shortly once HistropediaJS is published on npm