Fit biography events when filters applied
- Animated pan to fit all visible biography events once filter applied
- For simplicity and performance, we simply jump to the correct zoom
level, then animate the pan/scroll part of the sequence. Because the
filter panel is always open when you make the filter selection, the
initial change to the timeline is not noticeable.

Note: Not needed when removing filters, as the current timeline
zoom/position can remain the same in this case